requirements.txt
pytest~=8.3.5
fastapi~=0.115.12
langchain~=0.3.21
python-dotenv~=1.0.1
langchain-openai~=0.3.9
uvicorn~=0.34.0
pydantic~=2.10.6
requests~=2.32.3
bs4~=0.0.2
beautifulsoup4~=4.13.3
openai~=1.68.2
faiss-cpu~=1.10.0
langchain-community~=0.3.20